LeiosTxCache: reorganize modules
